Move resource-shard functions to cluster datamanager
